# project-notes/api-spec.yaml
openapi: 3.1.0
info:
  title: Project Allocation Management System (PAMS) API
  version: 1.12.0
  description: |
    REST API for the Project Allocation Management System.
    All endpoints require a Keycloak-issued OIDC Bearer token unless stated otherwise.
    Obtain tokens via the Keycloak Authorization Code flow at the configured realm (`/realms/pams`).
    Identity resolution uses the `empCode` JWT claim for Keycloak-to-PAMS identity bridging (sub claim verified against employees table with empCode fallback).
    Dates use ISO 8601 `YYYY-MM-DD` format. Timestamps use UTC (`YYYY-MM-DDTHH:MM:SSZ`).

    **Pagination (NFR-18/19):** All list endpoints support `page` (default 1) and `limit` (default 10, max 100).
    **Single-endpoint strategy (NFR-20):** Each resource has one list endpoint that also supports search and filter via query parameters.
    **Role-based response shaping (NFR-21):** Responses are filtered server-side based on the authenticated user's role.
    **Consolidated REST design:** Deactivation is handled via PUT (state change), allocation stop via PATCH, allocation remove via DELETE (soft-delete).
    **Billable flag & project status:** Projects carry a `billable` boolean (defaults to true for Client accounts) and a lifecycle `status` (Upcoming → Active → Completed). Both HR and PM (scoped to own projects) can modify status.
...
  contact:
    name: PAMS Engineering
  license:
    name: Internal Use Only
servers:
  - url: https://api.pams.internal/api/v1
...
security:
  - BearerAuth: []
...
paths:
...
  /projects:
    get:
      tags: [Projects]
      summary: List projects
      operationId: projects_list
      parameters:
        - $ref: "#/components/parameters/PageParam"
        - $ref: "#/components/parameters/LimitParam"
        - name: search
          in: query
          schema:
            type: string
          description: Search by project code or name (partial match, case-insensitive).
        - name: accountCode
          in: query
          schema:
            type: string
          description: Filter by account code.
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ProjectStatus"
        - name: isActive
          in: query
          schema:
            type: boolean
        - name: billable
          in: query
          schema:
            type: boolean
          description: Filter by billable flag.
        - name: projectManagerEmpCode
          in: query
          schema:
            type: string
          description: Filter projects where specified employee is PM. Use with current user's empCode to get "my projects".
        - $ref: "#/components/parameters/SortParam"
      responses:
        "200":
          description: Paginated list of projects
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PagedResponse_ProjectSummary"
        "400":
          $ref: "#/components/responses/InvalidSortError"
        "401":
          $ref: "#/components/responses/Unauthorized"

    post:
      tags: [Projects]
      summary: Create a new project (HR only)
      operationId: projects_create
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProjectRequest"
            example:
              projectCode: PROJ-ALPHA
              projectName: Project Alpha
              accountCode: ACC-ACME
              projectManagerEmpCode: EMP001
              billable: true
              startDate: "2026-03-01"
              endDate: "2026-12-31"
      responses:
        "201":
          description: Project created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          description: Project code already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProblemDetails"
  /projects/{projectCode}:
    parameters:
      - $ref: "#/components/parameters/ProjectCode"

    get:
      tags: [Projects]
      summary: Get project by code
      operationId: projects_getByCode
      responses:
        "200":
          description: Project detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"

    put:
      tags: [Projects]
      summary: Update project (HR or PM — PM scoped to own projects). Project code is immutable. Set isActive=false to deactivate.
      operationId: projects_update
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
      responses:
        "200":
          description: Project updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
  /projects/export:
    post:
      tags: [Projects]
      summary: Export projects as PDF or Excel (HR, PM)
      description: |
        Returns all matching projects (no pagination) as a downloadable PDF or Excel file.
        Accepts the same filter and sort query parameters as the list endpoint.
        An optional JSON body may specify a column name map to control which columns appear
        in the export and their display labels.
      operationId: projects_export
      parameters:
        - name: search
          in: query
          schema:
            type: string
          description: Search by project code or name.
        - name: accountCode
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            $ref: "#/components/schemas/ProjectStatus"
        - name: isActive
          in: query
          schema:
            type: boolean
        - name: billable
          in: query
          schema:
            type: boolean
        - name: projectManagerEmpCode
          in: query
          schema:
            type: string
        - $ref: "#/components/parameters/SortParam"
        - $ref: "#/components/parameters/ExportExtParam"
      requestBody:
        required: false
        description: |
          Optional column mapping. Keys are response DTO field names (case-insensitive),
          values are display labels for the export header. Only matched columns are included.
          If omitted or empty, all default columns are exported.
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ExportColumnMap"
            example:
              projectCode: "Project Code"
              projectName: "Project Name"
              accountName: "Account"
              status: "Status"
      responses:
        "200":
          description: File download
          content:
            application/pdf:
              schema:
                type: string
                format: binary
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
          headers:
            Content-Disposition:
              schema:
                type: string
              example: 'attachment; filename="projects-2026-03-05.pdf"'
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"

  # ── PROJECT TEAM MEMBERS (FR-020) ─────────────────────────────────────────
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        OIDC Bearer token issued by Keycloak (RS256 signed).
        Obtain via Keycloak Authorization Code flow: `{Authority}/protocol/openid-connect/auth`
        or Client Credentials flow for machine-to-machine.
        Default realm: `https://keycloak.internal/realms/pams`.

  # ── PARAMETERS ─────────────────────────────────────────────────────────────
  parameters:
...
    ProjectCode:
      name: projectCode
      in: path
      required: true
      schema:
        type: string
        maxLength: 50
      example: PROJ-ALPHA
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    LimitParam:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      description: "Number of items per page (default: 10, max: 100). Per NFR-18."
    SortParam:
      name: sort
      in: query
      schema:
        type: string
      description: |
        Sort field. Use `fieldName` for ascending or `-fieldName` for descending.
        Each endpoint has a default sort if omitted. Only one sort field per request.
        Invalid field names return 400 with a ProblemDetails response containing
        `field` (the invalid field) and `allowedFields` (list of valid field names).
      example: "-fromDate"
    ExportExtParam:
      name: ext
      in: query
      required: true
      schema:
        type: string
        enum: [pdf, xls]
      description: "Export format: `pdf` for PDF or `xls` for Excel (.xlsx)."

  # ── RESPONSES ──────────────────────────────────────────────────────────────
  responses:
    Unauthorized:
      description: Missing or invalid authentication token
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          example:
            type: "https://pams.internal/errors/ERR_UNAUTHORIZED"
            title: Unauthorized
            status: 401
            detail: Authentication token is missing or invalid.
    Forbidden:
      description: Insufficient permissions for this operation
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          example:
            type: "https://pams.internal/errors/ERR_FORBIDDEN"
            title: Forbidden
            status: 403
            detail: You do not have permission to perform this action.
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ProblemDetails"
          example:
            type: "https://pams.internal/errors/ERR_NOT_FOUND"
            title: Not Found
            status: 404
            detail: The requested resource was not found.
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ValidationProblemDetails"
          example:
            type: "https://pams.internal/errors/ERR_VALIDATION"
            title: Validation Failed
            status: 400
            errors:
              Percentage:
                - Percentage must be at least 25%.
                - Percentage must be a multiple of 5%.
...
    InvalidSortError:
      description: "Invalid sort field. Returns the field name and list of allowed sort fields."
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/InvalidSortProblem"
          example:
            type: "https://pams.internal/errors/ERR_INVALID_SORT"
            title: "Invalid Sort Parameter"
            status: 400
            detail: "Invalid sort field 'xyz'. Allowed fields: accountCode, accountName, accountType, isActive"
            instance: "/api/v1/accounts"
            field: "xyz"
            allowedFields: ["accountCode", "accountName", "accountType", "isActive"]

  # ── SCHEMAS ────────────────────────────────────────────────────────────────
  schemas:
...
    ProjectStatus:
      type: string
      enum: [Upcoming, Active, Completed]
      description: |
        Lifecycle stage of a project. Manually managed by HR or PM — not automatically derived from dates.
        Upcoming = project created but work has not started.
        Active = work is in progress.
        Completed = all work finished, project closed.
...
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          format: uri
        title:
          type: string
        status:
          type: integer
        detail:
          type: string
        instance:
          type: string
        traceId:
          type: string
...
    InvalidSortProblem:
      allOf:
        - $ref: "#/components/schemas/ProblemDetails"
        - type: object
          properties:
            field:
              type: string
              description: The invalid sort field that was provided.
            allowedFields:
              type: array
              items:
                type: string
              description: List of valid sort field names for this endpoint.
          example:
            type: "https://pams.internal/errors/ERR_INVALID_SORT"
            title: "Invalid Sort Parameter"
            status: 400
            detail: "Invalid sort field 'xyz'. Allowed fields: accountCode, accountName, accountType, isActive, totalActiveProjects, totalInactiveProjects"
            field: "xyz"
            allowedFields: ["accountCode", "accountName", "accountType", "isActive", "totalActiveProjects", "totalInactiveProjects"]
    ExportColumnMap:
      type: object
      nullable: true
      additionalProperties:
        type: string
      description: |
        Column name mapping for export. Keys are response DTO field names (case-insensitive match),
        values are the display labels to use in the export header. Only matched columns are included.
        If null or empty, all default columns are exported.
      example:
        empCode: "Employee Code"
        fullName: "Name"

    # ── Pagination ─────────────────────────────────────────────────────────
    PaginationMeta:
      type: object
      description: "Standard pagination metadata per NFR-18."
      properties:
        page:
          type: integer
          examples: [1]
        limit:
          type: integer
          examples: [10]
        totalRecords:
          type: integer
          examples: [125]
        totalPages:
          type: integer
          examples: [13]

    # ── Auth ───────────────────────────────────────────────────────────────
    # Login/logout handled by Keycloak directly. No LoginRequest/LoginResponse schemas needed.
    # Authenticated user profile is returned via GET /employees/me (EmployeeDetailResponse).
    # CurrentUserResponse is kept for lightweight "who am I" contexts where full detail is not needed.
...
    ProjectSummary:
      type: object
      properties:
        projectId:
          type: string
          format: uuid
        projectCode:
          type: string
        projectName:
          type: string
        accountId:
          type: string
          format: uuid
        accountCode:
          type: string
        accountName:
          type: string
        projectManagerId:
          type: ["string", "null"]
          format: uuid
        projectManagerEmpCode:
          type: ["string", "null"]
        projectManagerName:
          type: ["string", "null"]
        status:
          $ref: "#/components/schemas/ProjectStatus"
        billable:
          type: boolean
          description: Whether the project is billable. Defaults to true for client accounts.
        isActive:
          type: boolean
          description: "Soft-delete/archive flag. False means deactivated. Independent of status — a Completed project can still be active (visible)."
        startDate:
          type: string
          format: date
        endDate:
          type: ["string", "null"]
          format: date
        resourceCount:
          type: integer
          description: "Count of active non-deleted allocations on this project."
    ProjectResponse:
      allOf:
        - $ref: "#/components/schemas/ProjectSummary"
        - type: object
          properties:
            resourceCount:
              type: integer
              description: "Count of active non-deleted allocations on this project."
            allocations:
              type: array
              items:
                $ref: "#/components/schemas/AllocationDetailResponse"
              description: All active allocations for this project
            teamMembers:
              type: array
              items:
                $ref: "#/components/schemas/ProjectTeamMemberResponse"
              description: Team lead/reportee assignments
            createdAt:
              type: string
              format: date-time
            updatedAt:
              type: string
              format: date-time
    CreateProjectRequest:
      type: object
      required: [projectCode, projectName, accountCode, startDate]
      properties:
        projectCode:
          type: string
          maxLength: 50
        projectName:
          type: string
          maxLength: 150
        accountCode:
          type: string
          maxLength: 50
          description: Account code (must exist and be active).
        projectManagerEmpCode:
          type: ["string", "null"]
          maxLength: 50
          description: Employee code of the PM (must be an active employee with ProjectManager role).
        startDate:
          type: string
          format: date
        endDate:
          type: ["string", "null"]
          format: date
        billable:
          type: boolean
          description: |
            Whether the project is billable.
            Defaults to true when the linked account's type is 'Client'; false otherwise.
            HR can override during creation.
    UpdateProjectRequest:
      type: object
      required: [projectName, startDate, status, billable]
      properties:
        projectName:
          type: string
          maxLength: 150
        projectManagerEmpCode:
          type: ["string", "null"]
          maxLength: 50
          description: Employee code of the PM.
        startDate:
          type: string
          format: date
        endDate:
          type: ["string", "null"]
          format: date
        status:
          $ref: "#/components/schemas/ProjectStatus"
        billable:
          type: boolean
          description: Whether the project is billable.
        isActive:
          type: boolean
          description: "Set to false to deactivate/archive the project. Independent of status. Does not affect existing allocations."
    PagedResponse_ProjectSummary:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/ProjectSummary"
        pagination:
          $ref: "#/components/schemas/PaginationMeta"

    # ── Employee ───────────────────────────────────────────────────────────
...
    ProjectTeamMemberResponse:
      type: object
      description: A project-scoped Team Lead → Reportee assignment.
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        projectCode:
          type: string
        teamLeadId:
          type: string
          format: uuid
        teamLeadEmpCode:
          type: string
        teamLeadFullName:
          type: string
        reporteeId:
          type: string
          format: uuid
        reporteeEmpCode:
          type: string
        reporteeFullName:
          type: string
        createdAt:
          type: string
          format: date-time
...