# src/PAMS.API/Controllers/ProjectsController.cs
using MediatR;
using Microsoft.AspNetCore.Authorization;
using Microsoft.AspNetCore.Mvc;
using PAMS.Application.Commands.Projects;
using PAMS.Application.DTOs.Allocations;
using PAMS.Application.DTOs.Common;
using PAMS.Application.DTOs.Projects;
using PAMS.Application.DTOs.ProjectTeamMembers;
using PAMS.Application.Interfaces;
using PAMS.Domain.Enums;
using PAMS.Domain.Repositories;

namespace PAMS.API.Controllers;

/// <summary>
/// Project management endpoints. Projects belong to accounts and have allocations.
/// </summary>
[ApiController]
[Route("api/v1/projects")]
[Authorize]
[Produces("application/json")]
[Tags("Projects")]
public sealed class ProjectsController : ControllerBase
{
    private readonly IMediator _mediator;
    private readonly IProjectRepository _projectRepo;
    private readonly IExportService _exportService;

    public ProjectsController(IMediator mediator, IProjectRepository projectRepo, IExportService exportService)
    {
        _mediator = mediator;
        _projectRepo = projectRepo;
        _exportService = exportService;
    }

    /// <summary>
    /// List projects with optional search and filters.
    /// </summary>
    /// <remarks>
    /// Returns a paginated list of projects. Supports search by project code or name,
    /// and filtering by account code, status, active status, billable flag, and PM employee code.
    /// Use projectManagerEmpCode with the current user's empCode to get "my projects".
    /// </remarks>
    /// <param name="search">Search by project code or name (partial match, case-insensitive).</param>
    /// <param name="accountCode">Filter by account code.</param>
    /// <param name="status">Filter by project status (Upcoming, Active, Completed).</param>
    /// <param name="isActive">Filter by active status.</param>
    /// <param name="billable">Filter by billable flag.</param>
    /// <param name="projectManagerEmpCode">Filter projects where specified employee is PM.</param>
    /// <param name="page">Page number (default: 1).</param>
    /// <param name="limit">Items per page (default: 10, max: 100).</param>
    /// <param name="sort">Sort field. Prefix with - for descending (e.g. -projectName).</param>
    /// <param name="ct">Cancellation token.</param>
    /// <returns>Paginated list of project summaries.</returns>
    [HttpGet]
    [ProducesResponseType(typeof(PagedResponse<ProjectSummaryResponse>), StatusCodes.Status200OK)]
    [ProducesResponseType(StatusCodes.Status401Unauthorized)]
    public async Task<ActionResult<PagedResponse<ProjectSummaryResponse>>> List(
        [FromQuery] string? search,
        [FromQuery] string? accountCode,
        [FromQuery] ProjectStatus? status,
        [FromQuery] bool? isActive,
        [FromQuery] bool? billable,
        [FromQuery] string? projectManagerEmpCode,
        [FromQuery] int page = 1,
        [FromQuery] int limit = 10,
        [FromQuery] string? sort = null,
        CancellationToken ct = default)
    {
        limit = Math.Clamp(limit, 1, 100);
        page = Math.Max(page, 1);

        var projects = await _projectRepo.GetFilteredAsync(
            search, accountCode, status, isActive, billable, projectManagerEmpCode, page, limit, sort, ct);
        var totalRecords = await _projectRepo.GetFilteredCountAsync(
            search, accountCode, status, isActive, billable, projectManagerEmpCode, ct);

        var data = projects.Select(p =>
        {
            var today = DateOnly.FromDateTime(DateTime.Today);
            var resourceCount = (p.Allocations ?? [])
                .Count(a => a.DeletedAt == null && a.FromDate <= today && (a.ToDate == null || a.ToDate >= today));

            return new ProjectSummaryResponse
            {
                ProjectId = p.Id,
                ProjectCode = p.ProjectCode,
                ProjectName = p.ProjectName,
                AccountId = p.AccountId,
                AccountCode = p.Account?.AccountCode ?? string.Empty,
                AccountName = p.Account?.AccountName ?? string.Empty,
                ProjectManagerId = p.ProjectManagerId,
                ProjectManagerEmpCode = p.ProjectManager?.EmpCode,
                ProjectManagerName = p.ProjectManager is not null
                    ? $"{p.ProjectManager.FirstName} {p.ProjectManager.LastName}" : null,
                Status = p.Status,
                Billable = p.Billable,
                IsActive = p.IsActive,
                StartDate = p.StartDate,
                EndDate = p.EndDate,
                ResourceCount = resourceCount
            };
        }).ToList();

        return Ok(new PagedResponse<ProjectSummaryResponse>
        {
            Data = data,
            Pagination = PaginationMeta.Create(page, limit, totalRecords)
        });
    }

    /// <summary>
    /// Export projects as PDF or Excel.
    /// </summary>
    /// <remarks>
    /// Returns all matching projects (no pagination) as a downloadable file.
    /// Accepts the same filter and sort parameters as the list endpoint.
    /// An optional JSON body may contain a column name map where keys are field names
    /// and values are display labels. Only mapped columns appear in the export.
    /// If no body is sent, all default columns are included.
    /// </remarks>
    /// <param name="search">Search by project code or name (partial match, case-insensitive).</param>
    /// <param name="accountCode">Filter by account code.</param>
    /// <param name="status">Filter by project status (Upcoming, Active, Completed).</param>
    /// <param name="isActive">Filter by active status. Omit to return all.</param>
    /// <param name="billable">Filter by billable flag.</param>
    /// <param name="projectManagerEmpCode">Filter by project manager employee code.</param>
    /// <param name="sort">Sort field. Prefix with - for descending (e.g. -projectName).</param>
    /// <param name="ext">Export format: pdf or xls.</param>
    /// <param name="columns">Optional column name map. Keys = field names, values = display labels.</param>
    /// <param name="ct">Cancellation token.</param>
    /// <returns>File download (PDF or Excel).</returns>
    /// <response code="200">File download.</response>
    /// <response code="400">Invalid export format or invalid sort field.</response>
    /// <response code="401">Missing or invalid authentication token.</response>
    /// <response code="403">Insufficient permissions.</response>
    [HttpPost("export")]
    [Authorize(Policy = "CanAllocate")]
    [Produces("application/pdf", "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet")]
    [ProducesResponseType(StatusCodes.Status200OK)]
    [ProducesResponseType(StatusCodes.Status400BadRequest)]
    [ProducesResponseType(StatusCodes.Status401Unauthorized)]
    [ProducesResponseType(StatusCodes.Status403Forbidden)]
    public async Task<IActionResult> Export(
        [FromQuery] string? search,
        [FromQuery] string? accountCode,
        [FromQuery] ProjectStatus? status,
        [FromQuery] bool? isActive,
        [FromQuery] bool? billable,
        [FromQuery] string? projectManagerEmpCode,
        [FromQuery] string? sort,
        [FromQuery] string ext,
        [FromBody] Dictionary<string, string>? columns = null,
        CancellationToken ct = default)
    {
        var format = ext?.ToLowerInvariant();
        if (format != "pdf" && format != "xls")
            return Problem(
                detail: "ext must be 'pdf' or 'xls'.",
                statusCode: StatusCodes.Status400BadRequest,
                title: "Bad Request",
                type: "https://pams.internal/errors/ERR_VALIDATION");

        var projects = await _projectRepo.GetFilteredAllAsync(
            search, accountCode, status, isActive, billable, projectManagerEmpCode, sort, ct);

        var today = DateOnly.FromDateTime(DateTime.Today);
        var data = projects.Select(p =>
        {
            var resourceCount = (p.Allocations ?? [])
                .Count(a => a.DeletedAt == null && a.FromDate <= today && (a.ToDate == null || a.ToDate >= today));

            return new ProjectSummaryResponse
            {
                ProjectId = p.Id,
                ProjectCode = p.ProjectCode,
                ProjectName = p.ProjectName,
                AccountId = p.AccountId,
                AccountCode = p.Account?.AccountCode ?? string.Empty,
                AccountName = p.Account?.AccountName ?? string.Empty,
                ProjectManagerId = p.ProjectManagerId,
                ProjectManagerEmpCode = p.ProjectManager?.EmpCode,
                ProjectManagerName = p.ProjectManager is not null
                    ? $"{p.ProjectManager.FirstName} {p.ProjectManager.LastName}" : null,
                Status = p.Status,
                Billable = p.Billable,
                IsActive = p.IsActive,
                StartDate = p.StartDate,
                EndDate = p.EndDate,
                ResourceCount = resourceCount
            };
        }).ToList();

        var result = await _exportService.GenerateProjectsAsync(
            data, format == "xls" ? "xlsx" : format, columns, ct);

        return File(result.FileBytes, result.ContentType, result.FileName);
    }

    /// <summary>
    /// Get project by code.
    /// </summary>
    /// <remarks>
    /// Returns detailed project information including account and project manager details.
    /// </remarks>
    /// <param name="projectCode">Unique project business code.</param>
    /// <param name="ct">Cancellation token.</param>
    /// <returns>Project detail.</returns>
    [HttpGet("{projectCode}")]
    [ProducesResponseType(typeof(ProjectDetailResponse), StatusCodes.Status200OK)]
    [ProducesResponseType(StatusCodes.Status401Unauthorized)]
    [ProducesResponseType(StatusCodes.Status404NotFound)]
    public async Task<ActionResult<ProjectDetailResponse>> GetByCode(
        string projectCode, CancellationToken ct)
    {
        var p = await _projectRepo.GetByCodeAsync(projectCode, ct);
        if (p is null) return NotFound();

        var allocations = (p.Allocations ?? [])
            .Where(a => a.DeletedAt == null)
            .Select(a => AllocationDetailResponse.MapFrom(a, a.Employee, p))
            .ToList();

        var today = DateOnly.FromDateTime(DateTime.Today);
        var resourceCount = (p.Allocations ?? [])
            .Count(a => a.DeletedAt == null && a.FromDate <= today && (a.ToDate == null || a.ToDate >= today));

        var teamMembers = (p.TeamMembers ?? [])
            .Select(tm => new ProjectTeamMemberResponse
            {
                Id = tm.Id,
                ProjectId = tm.ProjectId,
                ProjectCode = p.ProjectCode,
                TeamLeadId = tm.TeamLeadId,
                TeamLeadEmpCode = tm.TeamLead?.EmpCode ?? string.Empty,
                TeamLeadFullName = tm.TeamLead is not null
                    ? $"{tm.TeamLead.FirstName} {tm.TeamLead.LastName}" : string.Empty,
                ReporteeId = tm.ReporteeId,
                ReporteeEmpCode = tm.Reportee?.EmpCode ?? string.Empty,
                ReporteeFullName = tm.Reportee is not null
                    ? $"{tm.Reportee.FirstName} {tm.Reportee.LastName}" : string.Empty,
                CreatedAt = tm.CreatedAt
            }).ToList();

        return Ok(new ProjectDetailResponse
        {
            ProjectId = p.Id,
            ProjectCode = p.ProjectCode,
            ProjectName = p.ProjectName,
            AccountId = p.AccountId,
            AccountCode = p.Account?.AccountCode ?? string.Empty,
            AccountName = p.Account?.AccountName ?? string.Empty,
            ProjectManagerId = p.ProjectManagerId,
            ProjectManagerEmpCode = p.ProjectManager?.EmpCode,
            ProjectManagerName = p.ProjectManager is not null
                ? $"{p.ProjectManager.FirstName} {p.ProjectManager.LastName}" : null,
            Status = p.Status,
            Billable = p.Billable,
            IsActive = p.IsActive,
            StartDate = p.StartDate,
            EndDate = p.EndDate,
            CreatedAt = p.CreatedAt,
            UpdatedAt = p.UpdatedAt,
            ResourceCount = resourceCount,
            Allocations = allocations,
            TeamMembers = teamMembers
        });
    }

    /// <summary>
    /// Create a new project (HR only).
    /// </summary>
    /// <remarks>
    /// Creates a project under an account with an assigned PM. Project code must be unique.
    /// Billable defaults to true for Client accounts, false otherwise.
    /// </remarks>
    /// <param name="command">Project creation request.</param>
    /// <param name="ct">Cancellation token.</param>
    /// <returns>Created project detail.</returns>
    [HttpPost]
    [Authorize(Policy = "HROnly")]
    [ProducesResponseType(typeof(ProjectDetailResponse), StatusCodes.Status201Created)]
    [ProducesResponseType(StatusCodes.Status400BadRequest)]
    [ProducesResponseType(StatusCodes.Status401Unauthorized)]
    [ProducesResponseType(StatusCodes.Status403Forbidden)]
    [ProducesResponseType(StatusCodes.Status409Conflict)]
    public async Task<ActionResult<ProjectDetailResponse>> Create(
        [FromBody] CreateProjectCommand command, CancellationToken ct)
    {
        var result = await _mediator.Send(command, ct);
        return CreatedAtAction(nameof(GetByCode),
            new { projectCode = result.ProjectCode }, result);
    }

    /// <summary>
    /// Update project (HR or PM — PM scoped to own projects).
    /// </summary>
    /// <remarks>
    /// Project code is immutable. Set isActive=false to deactivate.
    /// PM can only update projects they manage. Both HR and PM can modify status.
    /// </remarks>
    /// <param name="projectCode">Project code to update.</param>
    /// <param name="request">Update request body.</param>
    /// <param name="ct">Cancellation token.</param>
    /// <returns>Updated project detail.</returns>
    [HttpPut("{projectCode}")]
    [Authorize(Policy = "CanAllocate")]
    [ProducesResponseType(typeof(ProjectDetailResponse), StatusCodes.Status200OK)]
    [ProducesResponseType(StatusCodes.Status400BadRequest)]
    [ProducesResponseType(StatusCodes.Status401Unauthorized)]
    [ProducesResponseType(StatusCodes.Status403Forbidden)]
    [ProducesResponseType(StatusCodes.Status404NotFound)]
    public async Task<ActionResult<ProjectDetailResponse>> Update(
        string projectCode,
        [FromBody] UpdateProjectRequest request,
        CancellationToken ct)
    {
        var command = new UpdateProjectCommand
        {
            ProjectCode = projectCode,
            ProjectName = request.ProjectName,
            ProjectManagerEmpCode = request.ProjectManagerEmpCode,
            StartDate = request.StartDate,
            EndDate = request.EndDate,
            Status = request.Status,
            Billable = request.Billable,
            IsActive = request.IsActive
        };

        var result = await _mediator.Send(command, ct);
        return Ok(result);
    }
}

public sealed record UpdateProjectRequest
{
    public required string ProjectName { get; init; }
    public string? ProjectManagerEmpCode { get; init; }
    public required DateOnly StartDate { get; init; }
    public DateOnly? EndDate { get; init; }
    /// <summary>Lifecycle stage of the project (Upcoming, Active, Completed).</summary>
    public required ProjectStatus Status { get; init; }
    /// <summary>Whether the project is billable.</summary>
    public required bool Billable { get; init; }
    /// <summary>Set to false to deactivate/archive the project. Does not affect allocations.</summary>
    public bool? IsActive { get; init; }
}